Rough draft of template files for graph system.
I don't like this method. The more I try to think about it and work with
it the less I like it. I've sent you all a long email explaining my
reasons as to why I don't like it, and the modified system that I've
come up with that I think will work better. I've also given an example
of what that model will look like. I'm going to leave this alone,
however, and I'll make a fork of it and work on the new design model in
that fork. If you do want to use this, recognise that it is not complete
as a skeleton, and would still require some work.